# services/API_DOCUMENTATION.yaml
openapi: 3.0.3
info:
  title: Apollo Platform API
  description: |
    ## Apollo Criminal Investigation Platform - Comprehensive API Documentation

    The Apollo Platform provides a complete suite of APIs for criminal investigation,
    intelligence gathering, and threat analysis. This documentation covers all available
    endpoints across Node.js microservices and Python intelligence services.

    ### Service Architecture
    - **Node.js Microservices** (Port 3000-3010): Authentication, Operations, Intelligence, Search, Analytics, Notifications
    - **Python Intelligence Server** (Port 8000): OSINT, Blockchain, Facial Recognition, Voice Analysis, GEOINT

    ### Authentication Methods
    - **JWT Bearer Token**: Primary authentication for all protected endpoints
    - **API Key**: Alternative authentication via `X-API-Key` header
    - **MFA**: Multi-factor authentication for sensitive operations

    ### Rate Limiting
    - Free tier: 100 requests/hour
    - Pro tier: 1,000 requests/hour
    - Enterprise: 10,000 requests/hour

    ### Error Handling
    All errors follow a consistent format with error codes and messages.

  version: 2.0.0
  contact:
    name: Apollo Platform Team
    email: dev@apollo-platform.com
    url: https://apollo-platform.com/support
  license:
    name: Proprietary
    url: https://apollo-platform.com/license
  x-logo:
    url: https://apollo-platform.com/logo.png
    altText: Apollo Platform
servers:
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT token authentication
    apiKeyAuth:
      type: apiKey
      in: header
      name: X-API-Key
      description: API key authentication
  schemas:
...
    ApiKey:
      type: object
      properties:
        id:
          type: string
          format: uuid
        name:
          type: string
          example: "Production Integration"
        keyPrefix:
          type: string
          example: "apollo_abc..."
        scopes:
          type: array
          items:
            type: string
          example: ["investigations:read", "targets:read"]
        rateLimit:
          type: integer
          example: 1000
        expiresAt:
          type: string
          format: date-time
        lastUsedAt:
          type: string
          format: date-time
        isActive:
          type: boolean

    # ============ Operation Schema ============
...
paths:
...
  /api/auth/api-keys:
    get:
      summary: List API keys
      description: Get all API keys for the authenticated user
      tags: [API Keys]
      security:
        - bearerAuth: []
      responses:
        '200':
          description: List of API keys
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  data:
                    type: object
                    properties:
                      keys:
                        type: array
                        items:
                          $ref: '#/components/schemas/ApiKey'
                      totalCount:
                        type: integer
    post:
      summary: Create API key
      description: Create a new API key (requires MFA)
      tags: [API Keys]
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [name, scopes]
              properties:
                name:
                  type: string
                  maxLength: 255
                scopes:
                  type: array
                  items:
                    type: string
                preset:
                  type: string
                  enum: [readonly, analyst, operator, admin, webhook]
                rateLimit:
                  type: integer
                  minimum: 1
                  maximum: 10000
                rateLimitWindow:
                  type: integer
                  minimum: 1
                  maximum: 3600
                expiresIn:
                  type: integer
                  minimum: 1
                  maximum: 365
                  description: Expiration in days
                ipWhitelist:
                  type: array
                  items:
                    type: string
      responses:
        '201':
          description: API key created
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  message:
                    type: string
                  data:
                    type: object
                    properties:
                      id:
                        type: string
                      apiKey:
                        type: string
                        description: Only shown once!
                      keyPrefix:
                        type: string
                      name:
                        type: string
                      scopes:
                        type: array
                        items:
                          type: string
                      expiresAt:
                        type: string
  /api/auth/api-keys/scopes:
    get:
      summary: Get available scopes
      description: List available API key scopes and presets
      tags: [API Keys]
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Available scopes
  /api/auth/api-keys/{keyId}:
    get:
      summary: Get API key details
      description: Get details of a specific API key
      tags: [API Keys]
      security:
        - bearerAuth: []
      parameters:
        - name: keyId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: API key details
        '404':
          description: API key not found
    delete:
      summary: Revoke API key
      description: Permanently revoke an API key
      tags: [API Keys]
      security:
        - bearerAuth: []
      parameters:
        - name: keyId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                reason:
                  type: string
      responses:
        '200':
          description: API key revoked
  /api/auth/api-keys/{keyId}/rotate:
    post:
      summary: Rotate API key
      description: Generate new key value (requires MFA)
      tags: [API Keys]
      security:
        - bearerAuth: []
      parameters:
        - name: keyId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: API key rotated
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  data:
                    type: object
                    properties:
                      id:
                        type: string
                      newApiKey:
                        type: string
                      keyPrefix:
                        type: string
  /api/auth/api-keys/{keyId}/usage:
    get:
      summary: Get API key usage stats
      description: Get usage statistics for an API key
      tags: [API Keys]
      security:
        - bearerAuth: []
      parameters:
        - name: keyId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Usage statistics

  # ============================================================
  # OAUTH ENDPOINTS
  # ============================================================
...
security:
  - bearerAuth: []
  - apiKeyAuth: []